ff syncen

+++ Periode 1/Week 2/Opgave 6/Opgave 6/Program.cs
@@ -11,7 +11,7 @@ namespace Opgave_6
         static void Main(string[] args)
         {
             int[,] temp = new int[4, 7] { { 15, 1, 2, 0, 20, 5, 7 }, { 6, 3, 0, 8, 9, 15, 7 }, { 2, 3, 5, 7, 8, 9, 10 }, { 11, 13, 16, 13, 15, 11, 7 } };
-            int totaal = 0, ave = 0, count = 0;
+            double totaal = 0, ave = 0, count = 0;
             Console.WriteLine("Maand: April");
             for (int i = 0; i < 4; i++)
             {
